Title: 실수 입력받아 계산, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240801/실수 입력받아 계산/calculation-by-inputting-a-real-number.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-    // 여기에 코드를 작성해주세요.
-    int a, b, sum;
+    // 여기에 코드를 작성해주세요.  
+    double a, b, sum;
     cin >> a;
     cin >> b;
